requirements.txt
asyncio
autobahn
pyserial

